chore(refactor)
Refactor utilities to simplify driver integration
