+++ src/puppeteer.js
@@ -1,57 +1,23 @@
-const { AxePuppeteer } = require('@axe-core/puppeteer');
-const FS = require('fs');
-const PATH = require('path');
-const puppeteer = require('puppeteer');
-const { prettyPrintHtml } = require('./utils');
-
 /**
  * @see {https://github.com/puppeteer/puppeteer/blob/v5.2.1/docs/api.md|Puppeteer API Reference}
- * @see {@link https://www.deque.com/axe/core-documentation/api-documentation/|AXE API Reference}
-
- * @typedef Action
- * @property {String} action - An action name, e.g., click, blur, select
- * @property {Object} options - An object passed through a puppeteer action
- * @property {String} options.button - One of 'left', 'middle', 'right'; defaults to 'left'.
- * @property {Number} options.clickCount - Defaults to 1.
- * @property {Number} options.delay - Milliseconds to wait in action; defaults to 0.
- * @property {String[]} keys - An object passed through a puppeteer action
- * @property {String} target - The selector of the action target
- * @property {String|String[]} value - An object passed through a puppeteer action
- * @property {Object} waitFor - A selector to wait for after an action has taken place.
- * @property {Boolean} waitFor.hidden - Wait until node is hidden
- * @property {String} waitFor.target: - Selector identifying wait target
- * @property {Boolean} waitFor.visible - Wait until node becomes visible
- *
- * @typedef ClipRegion
- * @property {Number} ClipRegion.height
- * @property {Number} ClipRegion.width
- * @property {Number} ClipRegion.x - Coordinate of top-left corner of clip area.
- * @property {Number} ClipRegion.y - Coordinate of top-left corner of clip area.
- *
- * @typedef Harness
- * @property {AxePuppeteer} Harness.axe
- * @property {Puppeteer#Browser} Harness.browser
- * @property {Boolean} Harness.console - Send violations to the console. Default is true.
- * @property {String} Harness.metadata - Path used to save metadata information
- * @property {Puppeteer#Page} Harness.page
- * @property {Boolean} Harness.quiet - Turns off detail in results.
- * @property {String} Harness.save - A destination path and filename for the axe results.
- * @property {Boolean} Harness.summary - Send metadata to the console. Default is false.
- * @property {String} Harness.target - The CSS selector for the targeted element.
- * @property {Boolean} Harness.throws - Throw an Error on violations. Default is false.
- * @property {Boolean} Harness.verbose - Use AxePuppeteer#Violations object. Default is false.
- *
- * @typedef Screenshot
- * @property {ClipRegion} Screenshot.clip - An object which specifies clipping region of the page.
- * @property {String} Screenshot.encoding - One of 'base64' or 'binary'. Default is 'binary'.
- * @property {Boolean} Screenshot.fullPage - Screenshot the full scrollable page.
- * @property {Boolean} Screenshot.omitBackground - Hides white background to allow transparent pics.
- * @property {String} Screenshot.path - The file path to save the image to.
- * @property {Number} Screenshot.quality - The quality of the image, between 0-100.
- * @property {String} Screenshot.type - One of 'jpeg' or 'png'.
  */
 
-const Harness = {
+const { AxePuppeteer } = require('@axe-core/puppeteer');
+const DRIVER = require('puppeteer');
+const FS = require('fs');
+
+const {
+  analyze,
+  clipregion,
+  harness,
+  inrange,
+  mkdir,
+  prettyPrintHtml,
+  table,
+  target,
+} = require('./utils');
+
+let Harness = {
   axe: null,
   browser: null,
   console: true,
@@ -68,339 +34,6 @@ let Screenshot = {
   fullPage: true,
   omitBackground: false,
   encoding: 'binary',
-};
-
-let MetaData = {};
-let Violations = [];
-
-/**
- * @private
- * @name gblMkDir
- * @returns {undefined}
- * @param {String} path
- */
-const gblMkDir = (path) => {
-  if (path) {
-    const dir = PATH.dirname(path);
-
-    if (!FS.existsSync(dir)) {
-      FS.mkdirSync(dir, { recursive: true });
-    }
-  }
-};
-/**
- * @private
- * @name gblNoOp
- * @returns {undefined}
- */
-const gblNoOp = () => undefined;
-
-/**
- * @private
- * @name axeDocCss, axeDocIcons, axeDocTable
- * @returns {String}
- */
-const axeDocCss = `
-table {
-  border-collapse: collapse;
-  width: 100%;
-}
-tbody tr:nth-of-type(odd) {
-  background-color: #efefef;
-}
-tbody td:nth-of-type(n+2) {
-  text-align: center;
-}
-tbody td.impact {
-  text-align: left;
-  white-space: nowrap;
-}
-td, th {
-  padding: 0.25em;
-  vertical-align: top;
-}
-th:first-of-type {
-  text-align: left;
-}
-.flex {
-  align-items: center;
-  display: inline-flex;
-}
-.icon {
-  cursor: default;
-  display: inline-block;
-  font-size: 1.5em;
-  margin-right: 0.25em;
-  text-align: center;
-  user-select: none;
-  width: 1em;
-}`;
-const axeDocIcons = {
-  minor: '⚑',
-  moderate: '☹',
-  serious: '⚠',
-  critical: '☠',
-};
-const axeDocTable = () => `
-<table>
-  <thead>
-    <tr>
-      <th scope="col">Rule</th>
-      <th scope="col">Impact</th>
-    </tr>
-  </thead>
-  <tbody>${Violations.map((rule) => {
-    const { any = [], all = [] } = rule.nodes.pop() || {};
-
-    return `
-    <tr>
-      <td>
-        <a target="_blank" href="${rule.helpUrl}">
-          ${rule.help.replace(/</g, '&lt;').replace(/>/g, '&gt;')}
-        </a>
-
-        <div>${any && any.length > 0 ? `
-          <p>
-            <em>Fix any of the following</em>
-          </p>
-          <ul>${any.map((failure) => `
-            <li>
-              ${failure.message.replace(/</g, '&lt;').replace(/>/g, '&gt;')}
-              ${failure.relatedNodes.length > 0 ? `
-              <div>
-                <em>Related Nodes</em>:
-                <ul>${failure.relatedNodes.map((node) => `
-                  <li>
-                    ${node.target.join(' ').replace(/</g, '&lt;').replace(/>/g, '&gt;')}
-                  </li>`).join('')}
-                </ul>
-              </div>` : ''}
-            </li>`).join('')}
-          </ul>` : ''}
-          ${all && all.length > 0 ? `
-          <p>
-            <em>Fix all of the following</em>
-          </p>
-          <ul>${all.map((failure) => `
-            <li>
-              ${failure.message.replace(/</g, '&lt;').replace(/>/g, '&gt;')}
-              ${failure.relatedNodes.length > 0 ? `
-              <div>
-                <em>Related Nodes</em>:
-                <ul>${failure.relatedNodes.map((node) => `
-                  <li>
-                    ${node.target.join(' ').replace(/</g, '&lt;').replace(/>/g, '&gt;')}
-                  </li>`).join('')}
-                </ul>
-              </div>` : ''}
-            </li>`).join('')}
-          </ul>` : ''}
-        </div>
-      </td>
-      <td class="impact">
-        <span class="flex">
-          <span aria-hidden="true" class="icon">
-            ${axeDocIcons[rule.impact]}
-          </span>
-          <span>${rule.impact}</span>
-        </span>
-      </td>
-    </tr>`;
-  }).join('')}
-  </tbody>
-</table>`;
-
-/**
- * @private
- * @name axeHandle
- * @returns {undefined}
- * @param {String} message
- * @description Sends the results message to the console, or throws it as an Error
- *  according to Harness.throw
- */
-const axeHandle = (message) => {
-  if (Harness.save) {
-    const doc = `<!DOCTYPE html>
-<html lang="en-US">
-<head>
-<meta charset="utf-8">
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<title>Evaluation Results</title>
-<style type="text/css">
-${axeDocCss}
-</style>
-</head>
-<body>
-${axeDocTable(Violations)}
-</body>
-</html>
-`;
-
-    gblMkDir(Harness.save);
-
-    FS.writeFile(
-      Harness.save,
-      doc,
-      { encoding: 'utf8' },
-      gblNoOp
-    );
-  }
-  if (Harness.console) {
-    /* eslint-disable no-console */
-    console.log(`\n${message}`);
-    console.log('\n\n=======================================\n\n');
-    /* eslint-enable no-console */
-  }
-  if (Harness.throw) {
-    throw new Error(message);
-  }
-};
-/**
- * @private
- * @name axeHead
- * @returns {String}
- */
-const axeHead = () => {
-  if (Harness.verbose) {
-    return JSON.stringify(Violations, null, 2);
-  }
-
-  return Violations.map((violation) => {
-    const {
-      help,
-      impact,
-      nodes,
-    } = violation;
-    const details = `\n\n${nodes.map((node) => node.html).join('\n')}`;
-
-    return `${help}: ${impact.toUpperCase()}${!Harness.quiet ? details : ''}`;
-  }).join(!Harness.quiet ? '\n---------------------------------------\n\n' : '\n');
-};
-/**
- * @private
- * @name axeSortByImpact
- * @returns {Object[]}
- * @param {Object[]} violations
- * @description Sorts the AxeResults#violations array by impact, moving higher priority
- *  violations to lower array indices.
- */
-const axeSortByImpact = (violations = []) => {
-  const IMPACT = ['minor', 'moderate', 'serious', 'critical'];
-  const sortByImpact = (a, b) => IMPACT.indexOf(b.impact) - IMPACT.indexOf(a.impact);
-  return violations.map((violation) => {
-    const { nodes, ...remainder } = violation;
-    nodes.forEach((node) => {
-      node.all.sort(sortByImpact);
-      node.any.sort(sortByImpact);
-      node.none.sort(sortByImpact);
-    });
-    nodes.sort(sortByImpact);
-
-    return {
-      ...remainder,
-      nodes,
-    };
-  }).sort(sortByImpact);
-};
-/**
- * @private
- * @name clipRegion
- * @returns {ClipRegion}
- * @param {Object} region
- * @description Evaluates the provided region and returns a valid ClipRegion or undefined.
- */
-const clipRegion = (region = {}) => {
-  const required = ['height', 'width', 'x', 'y'];
-
-  // undefined values (missing required properties) are NaN
-  const nan = required.map((key) => Number.isNaN(Number(region[key]))).filter((v) => v);
-
-  return nan.length === 0 ? region : undefined;
-};
-/**
- * @private
- * @name getInRange
- * @returns {Number}
- * @param {Number|String} num
- * @param {Number[]} range
- * @description Evaluates the value to ensure it's in the specified range.
- */
-const getInRange = (value, range = [0, 0]) => {
-  const test = Number(value);
-
-  return !Number.isNaN(test) ? Math.min(Math.max(test, range[0]), range[1]) : undefined;
-};
-
-/**
- * @name analyze
- * @returns {Promise}
- * @param {Action} [action]
- * @param {Boolean} [override]
- * @description Runs AxePuppeteer#analyze and formats the response using the private functions,
- *  unless the handlers are overriden.
- */
-const analyze = async (action, override) => {
-  let results;
-
-  if (Harness.axe) {
-    try {
-      results = await Harness.axe.analyze();
-
-      const {
-        inapplicable,
-        incomplete,
-        passes,
-        testEngine,
-        testEnvironment,
-        timestamp,
-        url,
-        violations,
-      } = results;
-
-      MetaData = {
-        action,
-        inapplicable: inapplicable.reduce((acc, val) => acc + val.nodes.length, 0),
-        incomplete: incomplete.reduce((acc, val) => acc + val.nodes.length, 0),
-        passes: passes.reduce((acc, val) => acc + val.nodes.length, 0),
-        testEngine,
-        testEnvironment,
-        timestamp,
-        url,
-        violations: violations.reduce((acc, val) => acc + val.nodes.length, 0),
-      };
-
-      // sort violations
-      Violations = axeSortByImpact(violations);
-
-      if (Harness.metadata) {
-        const out = `${Harness.metadata}${PATH.sep}${timestamp.replace(/\D/g, '')}.json`;
-
-        gblMkDir(out);
-
-        FS.writeFile(
-          out,
-          JSON.stringify(MetaData),
-          { encoding: 'utf8' },
-          gblNoOp
-        );
-      }
-      if (Harness.summary) {
-        /* eslint-disable no-console */
-        console.log(`\n${JSON.stringify(MetaData, 2)}\n`);
-        /* eslint-enable no-console */
-      }
-
-      if (!override) {
-        if (Violations.length > 0) {
-          axeHandle(axeHead());
-        }
-      }
-    } catch (error) {
-      throw new Error(error);
-    }
-  }
-
-  return results;
 };
 
 /**
@@ -577,7 +210,7 @@ const html = async () => `
 <title>Evaluation Results</title>
 </head>
 <body>
-${axeDocTable(Violations)}
+${table()}
 </body>
 </html>
 `;
@@ -615,31 +248,39 @@ const launch = async (config = {}) => {
   } = config;
 
   try {
-    Harness.console = useConsole;
-    Harness.metadata = metadata;
-    Harness.quiet = quiet;
-    Harness.save = save;
-    Harness.summary = summary;
-    Harness.throws = throws;
-    Harness.verbose = verbose;
+    // initialize the harness
+    Harness = harness({
+      console: useConsole,
+      metadata,
+      quiet,
+      save,
+      summary,
+      throws,
+      verbose,
+    });
 
-    // make sure the results file is ready
-    if (FS.existsSync(Harness.save)) {
-      FS.unlinkSync(Harness.save);
-    }
-    if (Harness.save) {
-      gblMkDir(Harness.save);
-    }
-
-    Harness.browser = await puppeteer.launch({ product: 'chrome' });
-    Harness.page = await Harness.browser.newPage();
-    await Harness.page.setBypassCSP(true); // required for accessibility testing
+    const browser = await DRIVER.launch({ product: 'chrome' });
+    const page = await browser.newPage();
+    await page.setBypassCSP(true); // required for accessibility testing
 
     if (height && width) {
-      await Harness.page.setViewport({ height, width });
+      await page.setViewport({ height, width });
     }
 
-    Harness.axe = await new AxePuppeteer(Harness.page);
+    const axe = await new AxePuppeteer(page);
+
+    // set the drivers on the harness
+    Harness = harness({ ...Harness, browser, page, axe });
+
+    // make sure the results file is ready
+    if (Harness.save) {
+      if (FS.existsSync(Harness.save)) {
+        FS.unlinkSync(Harness.save);
+      }
+      if (Harness.save) {
+        mkdir(Harness.save);
+      }
+    }
 
     if (url) {
       await Harness.page.goto(url);
@@ -755,11 +396,11 @@ const setScreenshot = async (config = {}) => {
   const enumEncoding = ['binary', 'base64'];
   const enumType = ['png', 'jpeg'];
 
-  const clip = clipRegion(newClip);
+  const clip = clipregion(newClip);
   const encoding = enumEncoding[Math.max(enumEncoding.indexOf(newEncoding), 0)];
   const fullPage = typeof newFullPage === 'boolean' ? newFullPage : true;
   const omitBackground = typeof newOmitBackground === 'boolean' ? newOmitBackground : false;
-  const quality = getInRange(newQuality, [0, 100]);
+  const quality = inrange(newQuality, [0, 100]);
   const type = enumType[Math.max(enumType.indexOf(newType), 0)];
 
   Screenshot = {
@@ -789,7 +430,7 @@ const snapshot = async (filename = '') => {
   let content;
   let screenshot;
 
-  gblMkDir(path);
+  mkdir(path);
 
   if (Harness.target) {
     const clip = await Harness.page.$eval(Harness.target, (node) => {
@@ -828,24 +469,6 @@ const snapshot = async (filename = '') => {
   content = prettyPrintHtml(content).replace(/\\{2}/g, '');
 
   return { content, screenshot };
-};
-
-/**
- * @name target
- * @returns {undefined}
- * @param {Object} selectors
- * @property {String} include - CSS selector for the single element to evaluate.
- * @property {String[]} exclude - CSS selectors to ignore during accessibility evaluation.
- */
-const target = async ({ include, exclude } = {}) => {
-  try {
-    Harness.target = include;
-
-    Harness.axe.exclude(exclude.map((css) => [css]));
-    Harness.axe.include(Harness.target);
-  } catch (error) {
-    throw new Error('Unable to set target in axe.');
-  }
 };
 
 module.exports = {
